project_structure(drone):added the cloning command to the drone pipeline
deleted the cloning command from the dockerfile and added it to the pipeline.
That was necessary because when done in the build, it will mostly
be an outdated version of the code. just after a new "release" of the image it
would be good.so it must be cloned while doing the test.